placing table related samples to deid_table.py file to avoid complexity errors.
